+++ app/views/business_mailer/business_welcome_email.html.erb
@@ -1,0 +1,5 @@
+<body>
+  <p>Hey <%= @business.name %>,</p>
+  <p>I'm Joe the founder of <%= link_to "railsdevs", root_url %>. I saw that you added your business profile and wanted to reach out. If there's anything I can do to make your search easier let me know.</p>
+  <p>What are you looking for in a rails developer?</p>
+</body>
