--- app/views/welcome_mailer/business_welcome_email.html.erb
@@ -1,9 +1,0 @@
-<body>
-  <p>Hey <%= @business.name %>,</p>
-  <p>
-    I'm Joe the founder of <%= link_to "railsdevs", root_url %>. I saw that you
-    added your business profile and wanted to reach out. If there's anything I
-    can do to make your search easier let me know.
-  </p>
-  <p>What are you looking for in a Rails developer?</p>
-</body>
