--- app/views/welcome_mailer/business_welcome_email.text.erb
@@ -1,7 +1,0 @@
-Hey <%= @business.name %>,
-
-I'm Joe the founder of railsdevs. I saw that you
-added your business profile and wanted to reach out. If there's anything I
-can do to make your search easier let me know.
-
-What are you looking for in a Rails developer?
