--- app/views/welcome_mailer/developer_welcome_email.html.erb
@@ -1,6 +1,0 @@
-<p>Hey <%= @developer.name %>,</p>
-<p>
-  I'm Joe the founder of <%= link_to "railsdevs", root_url %>. I saw that you
-  added your developer profile and wanted to reach out.
-</p>
-<p>What does your ideal next role look like?</p>
