--- app/views/welcome_mailer/developer_welcome_email.text.erb
@@ -1,6 +1,0 @@
-Hey <%= @developer.name %>,
-
-I'm Joe the founder of railsdevs. I saw that you
-added your developer profile and wanted to reach out.
-
-What does your ideal next role look like?
