--- test/fixtures/welcome_mailer/welcome_business_text
@@ -1,7 +1,0 @@
-Hey Business Owner with Conversation,
-
-I'm Joe the founder of railsdevs. I saw that you
-added your business profile and wanted to reach out. If there's anything I
-can do to make your search easier let me know.
-
-What are you looking for in a Rails developer?
