--- test/fixtures/welcome_mailer/welcome_developer_text
@@ -1,6 +1,0 @@
-Hey Prospect Developer,
-
-I'm Joe the founder of railsdevs. I saw that you
-added your developer profile and wanted to reach out.
-
-What does your ideal next role look like?
