+++ test/mailers/previews/business_mailer_preview.rb
@@ -5,4 +5,8 @@ class BusinessMailerPreview < ActionMailer::Preview
 
     BusinessMailer.with(business:, developers:).developer_profiles
   end
+
+  def welcome_business
+    BusinessMailer.with(business: Business.first).business_welcome_email
+  end
 end
